refactor(ledger): collocate the segment route registrars :hammer:
Both segment registrars move out of segment_handler.go into
segment_routes.go, leaving the handler file with the handler struct, the
request/response envelopes and the cores.

No route, authz tuple, operation ID or path changed.
